requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.1.0
Django==3.2.8
django-celery-beat==2.2.1
django-timezone-field==4.2.1
djangorestframework==3.12.4
drf-yasg==1.20.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.3.0
idna==3.2
importlib-metadata==4.8.1
inflection==0.5.1
itypes==1.2.0
jieba3k==0.35.1
Jinja2==3.0.2
joblib==1.1.0
kombu==5.1.0
lxml==4.6.3
MarkupSafe==2.0.1
newspaper3k==0.2.8
nltk==3.6.5
packaging==21.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
pyparsing==2.4.7
python-crontab==2.6.0
python-dateutil==2.8.2
python-memcached==1.59
pytz==2021.3
PyYAML==5.4.1
redis==3.5.3
regex==2021.10.8
requests==2.26.0
requests-file==1.5.1
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
tinysegmenter==0.3
tldextract==3.1.2
tqdm==4.62.3
typing-extensions==3.10.0.2
uritemplate==4.0.0
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
